Roman To Integer
We utilize a hashmap to store the integer values corresponding to each Roman numeral character. By iterating through the characters of the input Roman numeral string, we convert each character into its integer representation. During this iteration, we check if the current character's value is less than the next character's value. If so, we subtract the current value from the result, adhering to the Roman numeral rule where smaller numerals before larger ones indicate subtraction. Otherwise, we add the current value to the result. This process accumulates the total integer value of the Roman numeral string, which is then returned as the final result.