Sort The Peoples Names
I have used HashMap in order to store the height as a key and names as an value then after storing it in HashMap then, we will sort the heights array. Then, traverse through height array but from end to start and store these value in the names  array from the starting index order. Then finally return the names array.