Find the Nth Root Of M
using binary search we are trying to find the nth root of the m by taking low and high variable as 1 and m respectively. Then running while loop in which if we get the number return it ,else if our value is < m then low = mid+1 , high = mid-1. if nothing is possible then simply return -1 . Also mention the edge case as if(m==0) return m;